Split shell polling tests
